Add new component TeamInfo, refactored out of MessageBoard
